+++ packages/components/Tabs/src/TabsItem.types.ts
@@ -5,7 +5,8 @@ import type { IViewProps } from '@fluentui-react-native/adapters';
 import { IFocusable, IPressableState } from '@fluentui-react-native/interactive-hooks';
 import { ITextProps } from '@fluentui-react-native/text';
 import { IconProps } from '@fluentui-react-native/icon';
-import { IButtonProps, IButtonTokens } from '@fluentui-react-native/button';
+import { FontTokens, IBackgroundColorTokens, IBorderTokens, IForegroundColorTokens } from '@fluentui-react-native/tokens';
+import { IPressableProps } from '@fluentui-react-native/pressable';
 
 export const tabsItemName = 'TabsItem';
 
@@ -35,7 +36,7 @@ export interface TabsItemState {
   info: TabsItemInfo;
 }
 
-export interface TabsItemProps extends IButtonProps {
+export interface TabsItemProps extends IPressableProps {
   /*
    ** The text string for the option
    */
@@ -62,22 +63,64 @@ export interface TabsItemProps extends IButtonProps {
    */
   accessibilityPosInSet?: number;
 
-   /*
-    ** Defines the number of tabs items in the group for accessibility purposes. It's recommended to set this value if tabs
-    ** items are not direct children of tabs. This value is auto-generated if tabs items are direct children of tabs.
-    */
+  /*
+   ** Defines the number of tabs items in the group for accessibility purposes. It's recommended to set this value if tabs
+   ** items are not direct children of tabs. This value is auto-generated if tabs items are direct children of tabs.
+   */
   accessibilitySetSize?: number;
+
+  /*
+   ** An accessibility label for narrator.
+   */
+  accessibilityLabel?: string;
 
   /**
    * A RefObject to access the IFocusable interface. Use this to access the public methods and properties of the component.
    */
   componentRef?: React.RefObject<IFocusable>;
+
+  /**
+   * Source URL or name of the icon to show on the TabsItem.
+   */
+  icon?: IconSourcesType;
+
+  testID?: string;
 }
 
-export interface TabsItemTokens extends IButtonTokens {
-  textBorderColor?: string;
+type IconSourcesType = number | string | IconProps;
+
+export interface TabsItemTokens extends IForegroundColorTokens, FontTokens, IBackgroundColorTokens, IBorderTokens {
+  /**
+   * The indicator color.
+   */
   indicatorColor?: string;
+
+  /**
+   * The icon color.
+   */
+  iconColor?: string;
+
+  /**
+   * Source URL or name of the icon to show on the TabsItem.
+   */
+  icon?: IconSourcesType;
+
+  /**
+   * Text to show on the TabsItem.
+   */
+  headerText?: string;
+
+  /**
+   * The amount of padding between the border and the headerText.
+   */
+  headerTextPadding?: number | string;
+
+  /**
+   * The amount of padding between the border and the headerText when the TabsItem has focus.
+   */
+  headerTextPaddingFocused?: number | string;
 }
+
 
 export interface TabsItemSlotProps {
   root: React.PropsWithRef<IViewProps>;