+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,64 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Mon, 09 Aug 2021 22:49:01 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Aug 2021 16:12:26 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.35.9
+
+Fri, 13 Aug 2021 16:12:26 GMT
+
+### Patches
+
+- Fix icons in experimental button and add fab (t-lindaweng@microsoft.com)
+
+## 0.35.8
+
+Fri, 13 Aug 2021 14:54:33 GMT
+
+### Patches
+
+- Add Expander component test (amdingler694@gmail.com)
+
+## 0.35.7
+
+Fri, 13 Aug 2021 09:05:57 GMT
+
+### Patches
+
+- Update apple native code and build flags (sanajmi@microsoft.com)
+
+## 0.35.6
+
+Thu, 12 Aug 2021 17:23:44 GMT
+
+### Patches
+
+- Patch up the tester (ruaraki@microsoft.com)
+
+## 0.35.5
+
+Wed, 11 Aug 2021 18:38:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/native-radio-button to v0.1.8 (chripear@microsoft.com)
+
+## 0.35.4
+
+Wed, 11 Aug 2021 04:24:12 GMT
+
+### Patches
+
+- Bump @fluentui/react-native to v0.24.33 (chripear@microsoft.com)
+
+## 0.35.3
+
+Tue, 10 Aug 2021 21:12:30 GMT
+
+### Patches
+
+- Patch tester (ruaraki@microsoft.com)
 
 ## 0.35.2