+++ packages/components/Stack/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/stack
 
-This log was last generated on Mon, 09 Aug 2021 22:49:01 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Aug 2021 16:12:26 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.25
+
+Fri, 13 Aug 2021 16:12:26 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/text to v0.10.8 (t-lindaweng@microsoft.com)
+
+## 0.5.24
+
+Thu, 12 Aug 2021 17:23:44 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.10.7 (ruaraki@microsoft.com)
+
+## 0.5.23
+
+Tue, 10 Aug 2021 21:12:30 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.10.6 (ruaraki@microsoft.com)
 
 ## 0.5.22