+++ packages/theming/apple-theme/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/apple-theme
 
-This log was last generated on Mon, 09 Aug 2021 22:49:01 GMT and should not be manually modified.
+This log was last generated on Fri, 13 Aug 2021 16:12:26 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.17
+
+Fri, 13 Aug 2021 16:12:26 GMT
+
+### Patches
+
+- Fix icons in experimental button and add fab (t-lindaweng@microsoft.com)
+
+## 0.7.16
+
+Thu, 12 Aug 2021 17:23:44 GMT
+
+### Patches
+
+- Rename tokens (ruaraki@microsoft.com)
+
+## 0.7.15
+
+Tue, 10 Aug 2021 21:12:30 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/default-theme to v0.7.9 (ruaraki@microsoft.com)
 
 ## 0.7.14